chore: applied suggestion
