+++ frontend/src/components/ScoreboardLeaderboard.tsx
@@ -150,7 +150,7 @@ const ScoreboardLeaderboard = ({ refreshTrigger = 0 }: ScoreboardLeaderboardProp
                                         style={{ gridTemplateColumns: gridTemplate(challenges.length) }}
                                         onClick={() => navigate(`/users/${entry.user_id}`)}
                                     >
-                                        <span className='text-xs text-text-subtle'>#{index + 1}</span>
+                                        <span className='text-xs text-text-subtle'>#{(pagination.page - 1) * pagination.page_size + index + 1}</span>
                                         <span className='text-xs font-semibold text-text'>{t('common.pointsShort', { points: entry.score })}</span>
                                         <div className='flex items-center gap-3.75 truncate'>
                                             <UserAvatar username={entry.username} size='sm' />
