+++ internal/repo/challenge_repo.go
@@ -75,6 +75,7 @@ func (r *ChallengeRepo) ListActiveFiltered(ctx context.Context, filter Challenge
 	}
 
 	offset := (page - 1) * pageSize
+	solveStatsJoin := "LEFT JOIN (SELECT s.challenge_id, COUNT(*) AS solve_count FROM submissions s JOIN users u ON u.id = s.user_id WHERE s.correct = true AND u.role != ? GROUP BY s.challenge_id) AS solve_stats ON solve_stats.challenge_id = challenge.id"
 	switch strings.TrimSpace(filter.Sort) {
 	case "", "latest":
 		listQuery = listQuery.Order("challenge.id DESC")
@@ -82,12 +83,12 @@ func (r *ChallengeRepo) ListActiveFiltered(ctx context.Context, filter Challenge
 		listQuery = listQuery.Order("challenge.id ASC")
 	case "most_solved":
 		listQuery = listQuery.
-			Join("LEFT JOIN (SELECT s.challenge_id, COUNT(*) AS solve_count FROM submissions s JOIN users u ON u.id = s.user_id WHERE s.correct = true AND u.role != 'blocked' GROUP BY s.challenge_id) AS solve_stats ON solve_stats.challenge_id = challenge.id").
+			Join(solveStatsJoin, models.BlockedRole).
 			OrderExpr("COALESCE(solve_stats.solve_count, 0) DESC").
 			Order("challenge.id DESC")
 	case "least_solved":
 		listQuery = listQuery.
-			Join("LEFT JOIN (SELECT s.challenge_id, COUNT(*) AS solve_count FROM submissions s JOIN users u ON u.id = s.user_id WHERE s.correct = true AND u.role != 'blocked' GROUP BY s.challenge_id) AS solve_stats ON solve_stats.challenge_id = challenge.id").
+			Join(solveStatsJoin, models.BlockedRole).
 			OrderExpr("COALESCE(solve_stats.solve_count, 0) ASC").
 			Order("challenge.id DESC")
 	default:
